#13 [FIX] 마이페이지 컴포넌트 분리 및 Zustand 적용
