+++ src/api/UserProfile.ts
@@ -1,7 +1,7 @@
-import { ApiResponse, ApiUpdateResponse } from '@/component/card/MyPageCard';
 import { fetchWithToken, getRequest } from './Request';
 import { UserData } from '@/components/ui/alarmbox';
 import { AddressResponse } from '@/component/ui/Address';
+import { ApiResponse, ApiUpdateResponse } from '@/interface/MyPageInterface';
 
 /**
  *
