+++ src/component/MyPage/AddressModal.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+function AddressModal() {
+  return <div>AddressModal</div>;
+}
+
+export default AddressModal;
